Redistribute ReaderStore methods to meet size limits
Split the 781-line ReaderStore core and oversized coordination
extensions into well-bounded files:

- ReaderStore+FolderWatchDispatch.swift  (84 lines) — auto-open event dispatching
- ReaderStore+Rendering.swift            (88 lines) — render pipeline and scheduling
- ReaderStore+Persistence.swift         (163 lines) — file I/O, save/load, logging
- ReaderStore+DocumentPresentation.swift(230 lines) — present/reload/route documents

Result: core file 398 lines (was 781), all coordination extensions
under 100 lines (DocumentOpenFlow 81, ExternalChangeFlow 91,
SourceEditing 88, TableOfContents 22).
