+++ minimark/Stores/Coordination/ReaderStore+DocumentOpenFlow.swift
@@ -50,65 +50,6 @@ extension ReaderStore {
         }
     }
 
-    private func applyPostOpenSideEffects(
-        accessibleURL: URL,
-        normalizedURL: URL,
-        origin: ReaderOpenOrigin,
-        initialDiffBaselineMarkdown: String?,
-        loadedMarkdown: String
-    ) {
-        let now = Date()
-        folderWatch.settler.beginSettling(
-            folderWatch.settler.makePendingContext(
-                origin: origin,
-                initialDiffBaselineMarkdown: initialDiffBaselineMarkdown,
-                loadedMarkdown: loadedMarkdown,
-                now: now
-            )
-        )
-        if let initialDiffBaselineMarkdown {
-            _ = diffBaselineTracker.recordAndSelectBaseline(
-                markdown: initialDiffBaselineMarkdown,
-                for: normalizedURL,
-                at: now
-            )
-        }
-        refreshOpenInApplications()
-        recordRecentManualOpenIfNeeded(accessibleURL, origin: origin)
-        notifyAutoLoadedFileIfNeeded(
-            normalizedURL,
-            origin: origin,
-            initialDiffBaselineMarkdown: initialDiffBaselineMarkdown
-        )
-        startWatchingCurrentFile()
-    }
-
-    private func recordRecentManualOpenIfNeeded(_ accessibleURL: URL, origin: ReaderOpenOrigin) {
-        guard origin == .manual else {
-            return
-        }
-
-        settingsStore.addRecentManuallyOpenedFile(accessibleURL)
-    }
-
-    private func notifyAutoLoadedFileIfNeeded(
-        _ normalizedURL: URL,
-        origin: ReaderOpenOrigin,
-        initialDiffBaselineMarkdown: String?
-    ) {
-        guard origin.shouldNotifyFileAutoLoaded,
-              activeFolderWatchSession != nil,
-              settingsStore.currentSettings.notificationsEnabled else {
-            return
-        }
-
-        folderWatch.systemNotifier.notifyFileChanged(
-            normalizedURL,
-            changeKind: initialDiffBaselineMarkdown == nil ? .added : .modified,
-            watchedFolderURL: activeFolderWatchSession?.folderURL
-        )
-    }
-
     func materializeDeferredDocument(
         origin: ReaderOpenOrigin? = nil,
         folderWatchSession: ReaderFolderWatchSession? = nil,
@@ -136,54 +77,5 @@ extension ReaderStore {
         if initialDiffBaselineMarkdown != nil {
             noteObservedExternalChange(kind: .modified)
         }
-    }
-
-    func handleIncomingOpenURL(_ url: URL) {
-        handleIncomingOpenURL(url, origin: .manual)
-    }
-
-    func handleIncomingOpenURL(
-        _ url: URL,
-        origin: ReaderOpenOrigin,
-        folderWatchSession: ReaderFolderWatchSession? = nil,
-        initialDiffBaselineMarkdown: String? = nil
-    ) {
-        guard url.isFileURL else {
-            return
-        }
-
-        guard Self.isSupportedMarkdownFileURL(url) else {
-            return
-        }
-
-        let normalizedIncomingURL = Self.normalizedFileURL(url)
-        if let fileURL, Self.normalizedFileURL(fileURL) == normalizedIncomingURL {
-            return
-        }
-
-        openFile(
-            at: normalizedIncomingURL,
-            origin: origin,
-            folderWatchSession: folderWatchSession,
-            initialDiffBaselineMarkdown: initialDiffBaselineMarkdown
-        )
-    }
-
-    func loadAndPresentDocument(
-        readURL: URL,
-        presentedAs fileURL: URL,
-        diffBaselineMarkdown: String?,
-        resetDocumentViewMode: Bool,
-        acknowledgeExternalChange: Bool
-    ) throws -> (markdown: String, modificationDate: Date) {
-        let loaded = try loadMarkdownFile(at: readURL)
-        try presentLoadedDocument(
-            loaded,
-            at: fileURL,
-            diffBaselineMarkdown: diffBaselineMarkdown,
-            resetDocumentViewMode: resetDocumentViewMode,
-            acknowledgeExternalChange: acknowledgeExternalChange
-        )
-        return loaded
     }
 }
