+++ minimark/Stores/Coordination/ReaderStore+FolderWatchDispatch.swift
@@ -1,0 +1,84 @@
+import Foundation
+
+extension ReaderStore {
+    func handleObservedWatchedFolderChanges(_ markdownFileEvents: [ReaderFolderWatchChangeEvent]) {
+        guard let session = activeFolderWatchSession else {
+            return
+        }
+
+        lastWatchedFolderEventAt = .now
+        let livePlan = liveFolderWatchAutoOpenPlan(for: markdownFileEvents, session: session)
+        updateFolderWatchAutoOpenWarning(livePlan.warning)
+        dispatchLiveFolderWatchAutoOpenEvents(
+            livePlan.autoOpenEvents,
+            session: session,
+            origin: .folderWatchAutoOpen
+        )
+    }
+
+    func openInitialMarkdownFilesFromWatchedFolder(
+        _ markdownFileEvents: [ReaderFolderWatchChangeEvent],
+        session: ReaderFolderWatchSession
+    ) {
+        folderWatchEventDispatchCoordinator.dispatchInitialEvents(
+            markdownFileEvents,
+            session: session
+        ) { [self] event, eventSession, eventOrigin in
+            openPrimaryFolderWatchAutoOpenEvent(
+                event,
+                session: eventSession,
+                origin: eventOrigin
+            )
+        }
+    }
+
+    // MARK: - Private Helpers
+
+    func liveFolderWatchAutoOpenPlan(
+        for markdownFileEvents: [ReaderFolderWatchChangeEvent],
+        session: ReaderFolderWatchSession
+    ) -> ReaderFolderWatchAutoOpenPlan {
+        folderWatch.autoOpenPlanner.livePlan(
+            for: markdownFileEvents,
+            activeSession: session,
+            currentDocumentFileURL: fileURLForCurrentDocument
+        )
+    }
+
+    func updateFolderWatchAutoOpenWarning(_ warning: ReaderFolderWatchAutoOpenWarning?) {
+        if let warning {
+            folderWatchAutoOpenWarning = warning
+        }
+    }
+
+    func dispatchLiveFolderWatchAutoOpenEvents(
+        _ plannedEvents: [ReaderFolderWatchChangeEvent],
+        session: ReaderFolderWatchSession,
+        origin: ReaderOpenOrigin
+    ) {
+        folderWatchEventDispatchCoordinator.dispatchLiveEvents(
+            plannedEvents,
+            session: session,
+            origin: origin
+        ) { [self] event, eventSession, eventOrigin in
+            openPrimaryFolderWatchAutoOpenEvent(
+                event,
+                session: eventSession,
+                origin: eventOrigin
+            )
+        }
+    }
+
+    func openPrimaryFolderWatchAutoOpenEvent(
+        _ event: ReaderFolderWatchChangeEvent,
+        session: ReaderFolderWatchSession,
+        origin: ReaderOpenOrigin
+    ) {
+        openFile(
+            at: event.fileURL,
+            origin: origin,
+            folderWatchSession: session,
+            initialDiffBaselineMarkdown: event.kind == .modified ? event.previousMarkdown : nil
+        )
+    }
+}
